refactor: 전역 로그인 체크 수정 - #114
